requirements.txt
argparse
colorama
termcolor
pathlib
anybase32
base36
base58
pybase62
base91
pybase100
exifread
opencv-python
pytesseract